@@ -12,10 +12,9 @@ description = "A project (and object) for storing, manipulating, and converting 
 readme = "README.md"
 license = { text = "MIT" }
 authors = [ { name = "Open Force Field Initiative", email = "info@openforcefield.org" } ]
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 classifiers = [
   "Programming Language :: Python :: 3 :: Only",
-  "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3.13",
   "Programming Language :: Python :: 3.14",
@@ -65,7 +64,7 @@ omit = [
 
 [tool.mypy]
 mypy_path = "stubs/"
-python_version = "3.11"
+python_version = "3.12"
 plugins = "pydantic.mypy"
 warn_unused_configs = true
 warn_unused_ignores = true
@@ -127,18 +126,14 @@ rdkit = ">=2024"
 openmm = "8.4.*"
 
 [tool.pixi.environments]
-py311amber = [ "py311", "test", "typing", "ambertools", "mosdef" ]
-py311openeye = [ "py311", "openeye", "test", "typing", "regression_tests", "engines" ]
-py311examples = [ "py311", "openeye", "test", "engines", "examples" ]
 py312amber = [ "py312", "test", "typing", "ambertools", "mosdef" ]
 py312openeye = [ "py312", "openeye", "test", "typing", "regression_tests", "engines" ]
 py312examples = [ "py312", "openeye", "test", "engines", "examples" ]
 py313openeye = [ "py313", "openeye", "test", "typing", "regression_tests", "engines" ]
 py313examples = [ "py313", "openeye", "test", "engines", "examples" ]
-py311etc = [ "py311", "openeye", "mosdef", "test", "typing", "etc" ]
 py312etc = [ "py312", "openeye", "mosdef", "test", "typing", "etc" ]
 dev = [ "py312", "openeye", "test", "typing", "dev", "engines" ]
-betas = [ "py311", "test", "typing", "engines", "betas" ]
+betas = [ "py312", "test", "typing", "engines", "betas" ]
 [tool.pixi.feature.ambertools.dependencies]
 ambertools = "24.*"
 numpy = "2.2.*"
@@ -181,8 +176,6 @@ gromacs = "*"
 [tool.pixi.feature.openeye.dependencies]
 openeye-toolkits = "*"
 
-[tool.pixi.feature.py311.dependencies]
-python = "3.11.*"
 [tool.pixi.feature.py312.dependencies]
 python = "3.12.*"
 [tool.pixi.feature.py313.dependencies]
