@@ -1,9 +1,8 @@
 from collections import defaultdict
-from typing import TYPE_CHECKING
 
-from openff.utilities.utilities import has_package, requires_package
+from openff.utilities import has_package, requires_package
 
-if TYPE_CHECKING or has_package("openmm"):
+if has_package("openmm"):
     import openmm
     import openmm.unit
 
