@@ -145,6 +145,6 @@ def test_hmr_with_ligand_virtual_sites(
 
     # make sure it can minimize and run briefly without crashing
     # this is a single ligand in vacuum with boring parameters
-    simulation.minimizeEnergy(maxIterations=100)
+    simulation.minimizeEnergy(maxIterations=10)
 
-    simulation.runForClockTime(10 * openmm.unit.second)
+    simulation.runForClockTime(2 * openmm.unit.second)
