@@ -6,7 +6,7 @@ from pathlib import Path
 from shutil import which
 
 from openff.toolkit import Quantity
-from openff.utilities.utilities import temporary_cd
+from openff.utilities import temporary_cd
 
 from openff.interchange import Interchange
 from openff.interchange.components.mdconfig import MDConfig
