@@ -1,6 +1,5 @@
 from abc import abstractmethod
 from copy import copy
-from typing import TYPE_CHECKING
 
 from openff.toolkit import Quantity, Topology
 
@@ -8,12 +7,11 @@ from openff.interchange.components.potentials import Collection, Potential
 from openff.interchange.foyer._guard import has_foyer
 from openff.interchange.models import PotentialKey, TopologyKey
 
-if TYPE_CHECKING:
-    if has_foyer:
-        try:
-            from foyer import Forcefield
-        except ModuleNotFoundError:
-            pass
+if has_foyer:
+    try:
+        from foyer import Forcefield
+    except ModuleNotFoundError:
+        pass
 
 
 # Is this the safest way to achieve PotentialKey id separation?
