@@ -2,15 +2,13 @@
 Helper functions for exporting the topology to OpenMM.
 """
 
-from typing import TYPE_CHECKING
-
 from openff.toolkit import Topology
 from openff.utilities import has_package
 
 from openff.interchange import Interchange
 from openff.interchange.models import BaseVirtualSiteKey
 
-if has_package("openmm") or TYPE_CHECKING:
+if has_package("openmm"):
     import openmm.app
 
 
